% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy.SDForest}
\alias{copy.SDForest}
\alias{copy}
\title{Copy a forest}
\usage{
\method{copy}{SDForest}(object, ...)
}
\arguments{
\item{object}{A SDForest object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A copy of the SDForest object
}
\description{
Returns a copy of the forest object.
Might be useful if you want to keep the original forest in comparison to the pruned forest.
}
\seealso{
\code{\link{prune}}
}
\author{
Markus Ulmer
}
