% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy.SDTree}
\alias{copy.SDTree}
\title{Copy a tree}
\usage{
\method{copy}{SDTree}(object, ...)
}
\arguments{
\item{object}{A SDTree object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A copy of the SDTree object
}
\description{
Returns a copy of the tree object. 
Might be useful if you want to keep the original tree in comparison to the pruned tree.
}
\seealso{
\code{\link{prune}}
}
\author{
Markus Ulmer
}
