% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDForest.r
\name{get_Q}
\alias{get_Q}
\title{Estimation of spectral transformation}
\usage{
get_Q(X, type, q_hat = 0)
}
\arguments{
\item{X}{numerical covariates of class \code{matrix}}

\item{type}{type of deconfounding, one of 'trim', 'DDL_trim', 'pca', 'no_deconfounding'}

\item{q_hat}{assumed confounding dimension, only needed for pca}
}
\value{
Q of class \code{matrix}, the spectral transformation
}
\description{
Estimates Q
\insertCite{Breiman1996BaggingPredictors}{SDForest}
}
\examples{
X <- matrix(rnorm(100 * 20), nrow = 100)
get_Q(X, 'trim')
get_Q(X, 'DDL_trim')
get_Q(X, 'pca', q_hat = 5)
get_Q(X, 'no_deconfounding')
}
\references{
\insertRef{Breiman1996BaggingPredictors}{SDForest}
}
