% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeForest.R
\name{mergeForest}
\alias{mergeForest}
\title{Merge two forests}
\usage{
mergeForest(fit1, fit2)
}
\arguments{
\item{fit1}{first \code{SDForest} object}

\item{fit2}{second \code{SDForest} object}
}
\value{
merged \code{SDForest} object
}
\description{
This function merges two forests. 
The trees are combined and the variable importance is 
calculated as a weighted average of the two forests. 
If the forests are trained on the same data, 
the predictions and oob_predictions are combined as well.
}
\author{
Markus Ulmer
}
