% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SDTree}
\alias{plot.SDTree}
\title{Plot SDTree}
\usage{
\method{plot}{SDTree}(object)
}
\arguments{
\item{object}{Fitted object of class \code{SDTree}.}
}
\description{
Plot the SDTree.
}
\seealso{
\code{\link{SDTree}}
}
\author{
Markus Ulmer
}
