% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{plotOOB}
\alias{plotOOB}
\title{Visualize the out-of-bag performance of an SDForest}
\usage{
plotOOB(object)
}
\arguments{
\item{object}{A paths object with loss_path \code{matrix} 
with the out-of-bag performance for each complexity parameter.}
}
\value{
A ggplot object
}
\description{
This function visualizes the out-of-bag performance of an SDForest
for different complexity parameters. Can be used to choose the optimal
complexity parameter.
}
\seealso{
\code{\link{regPath.SDForest}}
}
\author{
Markus Ulmer
}
