% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.SDForest}
\alias{predict.SDForest}
\title{Predictions for the SDForest}
\usage{
\method{predict}{SDForest}(object, newdata)
}
\arguments{
\item{object}{Fitted object of class \code{SDForest}.}

\item{newdata}{New test data of class \code{data.frame} containing
the covariates for which to predict the response.}
}
\value{
A vector of predictions for the new data.
}
\description{
Predicts the response for new data using a fitted SDForest.
}
\examples{
set.seed(1)
n <- 50
X <- matrix(rnorm(n * 20), nrow = n)
y <- sign(X[, 1]) * 3 + rnorm(n)
model <- SDForest(x = X, y = y, Q_type = 'no_deconfounding', nTree = 10)
predict(model, newdata = data.frame(X))
}
\seealso{
\code{\link{SDForest}}
}
\author{
Markus Ulmer
}
