% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDForest.r
\name{print.SDTree}
\alias{print.SDTree}
\title{Print Spectral Deconfounded Tree}
\usage{
print.SDTree(object)
}
\arguments{
\item{object}{Fitted object of class \code{SDTree}.}
}
\description{
Print contents of the spectral deconfounded tree.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{SDTree}}
}
\author{
Markus Ulmer
}
