% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDForest.R
\name{SDForest}
\alias{SDForest}
\title{Spectral Deconfounded Random Forest}
\usage{
SDForest(
  formula = NULL,
  data = NULL,
  x = NULL,
  y = NULL,
  nTree = 100,
  cp = 0,
  min_sample = 5,
  mtry = NULL,
  mc.cores = 1,
  Q_type = "trim",
  trim_quantile = 0.5,
  q_hat = 0,
  Q = NULL,
  A = NULL,
  gamma = 0.5,
  max_size = NULL,
  gpu = FALSE,
  return_data = TRUE,
  mem_size = 1e+07,
  leave_out_ind = NULL,
  envs = NULL,
  leave_envs_out_trees = NULL,
  envs_trees = NULL,
  max_candidates = 100
)
}
\arguments{
\item{formula}{Object of class \code{formula} or describing the model to fit 
of the form \code{y ~ x1 + x2 + ...} where \code{y} is a numeric response and 
\code{x1, x2, ...} are vectors of covariates. Interactions are not supported.}

\item{data}{Training data of class \code{data.frame} containing the variables in the model.}

\item{x}{Predictor data, alternative to \code{formula} and \code{data}.}

\item{y}{Response vector, alternative to \code{formula} and \code{data}.}

\item{nTree}{Number of trees to grow.}

\item{cp}{Complexity parameter, minimum loss decrease to split a node. 
A split is only performed if the loss decrease is larger than \code{cp * initial_loss}, 
where \code{initial_loss} is the loss of the initial estimate using only a stump.}

\item{min_sample}{Minimum number of observations per leaf. 
A split is only performed if both resulting leaves have at least 
\code{min_sample} observations.}

\item{mtry}{Number of randomly selected covariates to consider for a split, 
if \code{NULL} all covariates are available for each split.}

\item{mc.cores}{Number of cores to use for parallel processing,
if \code{mc.cores > 1} the trees are estimated in parallel.}

\item{Q_type}{Type of deconfounding, one of 'trim', 'pca', 'no_deconfounding'. 
'trim' corresponds to the Trim transform \insertCite{Cevid2020SpectralModels}{SDForest} 
as implemented in the Doubly debiased lasso \insertCite{Guo2022DoublyConfounding}{SDForest}, 
'pca' to the PCA transformation\insertCite{Paul2008PreconditioningProblems}{SDForest}. 
See \code{\link{get_Q}}.}

\item{trim_quantile}{Quantile for Trim transform, 
only needed for trim and DDL_trim, see \code{\link{get_Q}}.}

\item{q_hat}{Assumed confounding dimension, only needed for pca, 
see \code{\link{get_Q}}.}

\item{Q}{Spectral transformation, if \code{NULL} 
it is internally estimated using \code{\link{get_Q}}.}

\item{A}{Numerical Anchor of class \code{matrix}. See \code{\link{get_W}}.}

\item{gamma}{Strength of distributional robustness, \eqn{\gamma \in [0, \infty]}. 
See \code{\link{get_W}}.}

\item{max_size}{Maximum number of observations used for a bootstrap sample.
If \code{NULL} n samples with replacement are drawn.}

\item{gpu}{If \code{TRUE}, the calculations are performed on the GPU. 
If it is properly set up.}

\item{return_data}{If \code{TRUE}, the training data is returned in the output.
This is needed for \code{\link{prune.SDForest}}, \code{\link{regPath.SDForest}}, 
and for \code{\link{mergeForest}}.}

\item{mem_size}{Amount of split candidates that can be evaluated at once.
This is a trade-off between memory and speed can be decreased if either
the memory is not sufficient or the gpu is to small.}

\item{leave_out_ind}{Indices of observations that should not be used for training.}

\item{envs}{Vector of environments which can be used for stratified tree fitting.}

\item{leave_envs_out_trees}{Number of trees that should be estimated while leaving
one of the environments out. Results in number of environments times number of trees.
NOT SUPPORTED YET}

\item{envs_trees}{Number of trees that should be estimated for each environment.
Results in number of environments times number of trees.
NOT SUPPORTED YET}

\item{max_candidates}{Maximum number of split points that are 
proposed at each node for each covariate.}
}
\value{
Object of class \code{SDForest} containing:
\item{predictions}{Vector of predictions for each observation.}
\item{forest}{List of SDTree objects.}
\item{var_names}{Names of the covariates.}
\item{oob_loss}{Out-of-bag loss. MSE}
\item{oob_SDloss}{Out-of-bag loss using the spectral transformation.}
\item{var_importance}{Variable importance.}
\item{oob_ind}{List of indices of trees that did not contain the observation in the training set.}
\item{oob_predictions}{Out-of-bag predictions.}
If \code{return_data} is \code{TRUE} the following are also returned:
\item{X}{Matrix of covariates.}
\item{Y}{Vector of responses.}
\item{Q}{Spectral transformation matrix.}
}
\description{
Estimate regression Random Forest using spectral deconfounding.
# TODO: add more details
}
\examples{
# TODO: add example
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{get_Q}}, \code{\link{get_W}}, \code{\link{SDTree}}
}
\author{
Markus Ulmer
}
