% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List.R
\name{toList.SDForest}
\alias{toList.SDForest}
\alias{toList}
\title{SDForest toList method}
\usage{
\method{toList}{SDForest}(object, ...)
}
\arguments{
\item{object}{an SDForest object with the trees in Node format}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
an SDForest object with the trees in list format
}
\description{
Converts the trees in an SDForest object from
class \code{Node} \insertCite{Glur2023Data.tree:Structure}{SDForest} to class \code{list}.
This makes it substantially easier to save the forest to disk.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fromList}} \code{\link{toList.SDTree}}
}
\author{
Markus Ulmer
}
