% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List.R
\name{toList.SDTree}
\alias{toList.SDTree}
\title{SDTree toList method}
\usage{
\method{toList}{SDTree}(object, ...)
}
\arguments{
\item{object}{an SDTree object with the tree in Node format}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
an SDTree object with the tree in list format
}
\description{
Converts the tree in an SDTree object from 
class \code{Node} \insertCite{Glur2023Data.tree:Structure}{SDForest} to class \code{list}.
This makes it substantially easier to save the tree to disk.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fromList}}
}
\author{
Markus Ulmer
}
