% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{varImp.SDForest}
\alias{varImp.SDForest}
\alias{varImp}
\title{Extract variable importance of a SDForest}
\usage{
\method{varImp}{SDForest}(object)
}
\arguments{
\item{object}{A SDForest object}
}
\value{
A named vector of variable importance
}
\description{
This function extracts the variable importance of a SDForest.
The variable importance is calculated as the mean of the variable importance
of all trees in the forest.
}
\examples{
data(iris)
fit <- SDForest(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, 
                 iris, nTree = 10)
varImp(fit)
}
\seealso{
\code{\link{varImp.SDTree}} \code{\link{SDForest}}
}
\author{
Markus Ulmer
}
