% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDForest.r
\name{get_Q}
\alias{get_Q}
\title{Estimation of spectral transformation}
\usage{
get_Q(X, type, trim_quantile = 0.5, confounding_dim = 0)
}
\arguments{
\item{X}{numerical covariates of class \code{matrix}}

\item{type}{type of deconfounding, one of 'trim', 'DDL_trim', 'pca', 'no_deconfounding'. 'trim' corresponds to the Trim transform \insertCite{Cevid2020SpectralModels}{SDForest}, 'DDL_trim' to the implementation of the Doubly debiased lasso \insertCite{Guo2022DoublyConfounding}{SDForest}, 'pca' to the PCA transformation\insertCite{Paul2008PreconditioningProblems}{SDForest} and 'no_deconfounding' to the Identity}

\item{trim_quantile}{quantile for Trim transform and DDL Trim transform, only needed for trim and DDL_trim}

\item{confounding_dim}{assumed confounding dimension, only needed for pca}
}
\value{
Q of class \code{matrix}, the spectral transformation
}
\description{
Estimates the spectral transformation Q for spectral deconfounding by shrinking the leading singular values of the covariates.
}
\examples{
X <- matrix(rnorm(100 * 20), nrow = 100)
get_Q(X, 'trim')
get_Q(X, 'DDL_trim')
get_Q(X, 'pca', q_hat = 5)
get_Q(X, 'no_deconfounding')
}
\references{
\insertAllCited{}
}
