% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDForest.r
\name{plot.SDTree}
\alias{plot.SDTree}
\title{Plot Spectral Deconfounded Tree}
\usage{
\method{plot}{SDTree}(object)
}
\arguments{
\item{object}{Fitted object of class \code{SDTree}.}
}
\description{
Plot the spectral deconfounded tree.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{SDTree}}
}
\author{
Markus Ulmer
}
