% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SDTree}
\alias{plot.SDTree}
\title{Plot SDTree}
\usage{
\method{plot}{SDTree}(x, ...)
}
\arguments{
\item{x}{Fitted object of class \code{SDTree}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot the SDTree.
}
\seealso{
\code{\link{SDTree}}
}
\author{
Markus Ulmer
}
