% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDForest.r
\name{predict.SDTree}
\alias{predict.SDTree}
\title{Predictions for the Spectral Deconfounded Tree}
\usage{
\method{predict}{SDTree}(object, newdata)
}
\arguments{
\item{object}{Fitted object of class \code{SDTree}.}

\item{newdata}{New test data of class \code{data.frame} containing the covariates for which to predict the response.}
}
\value{
A vector of predictions for the new data.
}
\description{
Predicts the response for new data using a fitted spectral deconfounded tree.
}
\examples{
# TODO: add example
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{SDTree}}
}
