% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SDForest}
\alias{print.SDForest}
\title{Print SDForest}
\usage{
print.SDForest(object)
}
\arguments{
\item{object}{Fitted object of class \code{SDForest}.}
}
\description{
Print contents of the SDForest.
}
\seealso{
\code{\link{SDForest}}
}
\author{
Markus Ulmer
}
