% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SDTree}
\alias{print.SDTree}
\title{Print a SDTree}
\usage{
\method{print}{SDTree}(x, ...)
}
\arguments{
\item{x}{Fitted object of class \code{SDTree}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of the SDTree.
}
\seealso{
\code{\link{SDTree}}
}
\author{
Markus Ulmer
}
