% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SDTree}
\alias{print.SDTree}
\title{Print a SDTree}
\usage{
print.SDTree(object)
}
\arguments{
\item{object}{Fitted object of class \code{SDTree}.}
}
\description{
Print contents of the SDTree.
}
\seealso{
\code{\link{SDTree}}
}
\author{
Markus Ulmer
}
