% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDForest.r
\name{SDTree}
\alias{SDTree}
\title{Spectral Deconfounded Tree}
\usage{
SDTree(
  formula = NULL,
  data = NULL,
  x = NULL,
  y = NULL,
  max_leaves = 50,
  cp = 0.01,
  min_sample = 5,
  mtry = NULL,
  fast = TRUE,
  multicore = F,
  mc.cores = NULL,
  Q_type = "trim",
  trim_quantile = 0.5,
  confounding_dim = 0,
  Q = NULL
)
}
\arguments{
\item{formula}{Object of class \code{formula} or describing the model to fit of the form \code{y ~ x1 + x2 + ...} where \code{y} is a numeric response and \code{x1, x2, ...} are vectors of covariates. Interactions are not supported.}

\item{data}{Training data of class \code{data.frame} containing the variables in the model.}

\item{x}{Predictor data, alternative to \code{formula} and \code{data}.}

\item{y}{Response vector, alternative to \code{formula} and \code{data}.}

\item{max_leaves}{Maximum number of leaves for the grown tree.}

\item{cp}{Complexity parameter, minimum loss decrease to split a node. A split is only performed if the loss decrease is larger than \code{cp * initial_loss}, where \code{initial_loss} is the loss of the initial estimate using only a stump.}

\item{min_sample}{Minimum number of observations per leaf. A split is only performed if both resulting leaves have at least \code{min_sample} observations.}

\item{mtry}{Number of randomly selected covariates to consider for a split, if \code{NULL} all covariates are available for each split.}

\item{fast}{If \code{TRUE}, only the optimal splitts in the new leaves are evaluated and the previously optimal splitts and their potential loss-decrease are reused. If \code{FALSE} all possible splitts in all the leaves are reevaluated after every split.}

\item{multicore}{If \code{TRUE} the optional splitts are evaluated in parallel (only available on unix systems).}

\item{mc.cores}{Number of cores to use for parallel computing, if \code{NULL} all available cores - 1 are used.}

\item{Q_type}{Type of deconfounding, one of 'trim', 'DDL_trim', 'pca', 'no_deconfounding'. 'trim' corresponds to the Trim transform \insertCite{Cevid2020SpectralModels}{SDForest}, 'DDL_trim' to the implementation of the Doubly debiased lasso \insertCite{Guo2022DoublyConfounding}{SDForest}, 'pca' to the PCA transformation\insertCite{Paul2008PreconditioningProblems}{SDForest} and 'no_deconfounding' to the Identity. See \code{\link{get_Q}}.}

\item{trim_quantile}{Quantile for Trim transform and DDL Trim transform, only needed for trim and DDL_trim, see \code{\link{get_Q}}.}

\item{confounding_dim}{Assumed confounding dimension, only needed for pca, see \code{\link{get_Q}}.}

\item{Q}{Spectral transformation, if \code{NULL} it is internally estimated using \code{\link{get_Q}}.}
}
\value{
Object of class \code{SDTree} containing
\item{predictions}{Predictions for the training set.}
\item{tree}{The estimated tree of type \code{data.tree} \insertCite{data.tree}{SDForest}. The tree contains the information about all the splits and the resulting estimates.}
\item{var_names}{Names of the covariates in the training data.}
}
\description{
Estimates a tree using spectral deconfounding. # TODO: add more details
}
\examples{
# TODO: add example
}
\references{
\insertAllCited{}
}
